pyproject.toml
[project]
name = "foundry-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "urwid",
    "anyio",
    "python-dotenv",
    "azure-identity",
    "azure-ai-projects",
]
